Add logic to invoke clean client side libraries command in test cleanup phase
