+++ test/LibraryManager.IntegrationTest/TestSolution/TestProjectCore20/libman.json
@@ -1,8 +1,6 @@
 {
   "version": "1.0",
   "libraries": [
-    {
-      
-    }
+
   ]
 }
