+++ test/LibraryManager.IntegrationTest/VisualStudioLibmanHostTests.cs
@@ -2,9 +2,10 @@
 using System.Diagnostics;
 using System.IO;
 using System.Runtime.InteropServices;
-using System.Threading;
+using Microsoft.Test.Apex.Services;
 using Microsoft.Test.Apex.VisualStudio;
 using Microsoft.Test.Apex.VisualStudio.Editor;
+using Microsoft.Test.Apex.VisualStudio.Shell;
 using Microsoft.Test.Apex.VisualStudio.Shell.ToolWindows;
 using Microsoft.Test.Apex.VisualStudio.Solution;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -17,9 +18,11 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
     public class VisualStudioLibmanHostTest : VisualStudioHostTest
     {
         // Solution consts
-        const string _testSolutionName = @"TestSolution.sln";
+        const string _libman = "libman.json";
+        const string _projectName = @"TestProjectCore20";
         const string _rootDirectoryName = @"TestSolution";
-
+        const string _testSolutionName = @"TestSolution.sln";
+        private string _initialLibmanFileContent;
         private static VisualStudioLibmanHostTest _instance;
         private static string _resultPath;
         private static string _solutionPath;
@@ -38,11 +41,49 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
 
             Solution.Open(_solutionPath);
             Solution.WaitForFullyLoaded(); // This will get modified after bug 627108 get fixed
+
+            string pathToLibmanFile = Path.Combine(SolutionRootPath, _projectName, _libman);
+            _initialLibmanFileContent = File.ReadAllText(pathToLibmanFile);
         }
 
         protected override void DoHostTestCleanup()
         {
+            ProjectTestExtension webProject = Solution[_projectName];
+            ProjectItemTestExtension libmanConfig = webProject[_libman];
+
+            if (libmanConfig != null)
+            {
+                CleanClientSideLibraries();
+
+                libmanConfig.Open();
+                Editor.Selection.SelectAll();
+                Editor.KeyboardCommands.Delete();
+                Editor.Edit.InsertTextInBuffer(_initialLibmanFileContent);
+
+                libmanConfig.Save();
+            }
+
             base.DoHostTestCleanup();
+        }
+
+        private void CleanClientSideLibraries()
+        {
+            Guid guid = Guid.Parse("44ee7bda-abda-486e-a5fe-4dd3f4cefac1");
+            uint commandId = 0x0200;
+            SolutionExplorerItemTestExtension libmanConfigNode = SolutionExplorer.FindItemRecursive(_libman);
+
+            if (libmanConfigNode != null)
+            {
+                libmanConfigNode.Select();
+
+                WaitFor.IsTrue(() =>
+                {
+                    CommandQueryResult queryResult = VisualStudio.ObjectModel.Commanding.QueryStatusCommand(guid, commandId);
+                    return queryResult.IsEnabled;
+                }, TimeSpan.FromMilliseconds(40000), TimeSpan.FromMilliseconds(500));
+
+                VisualStudio.ObjectModel.Commanding.ExecuteCommand(guid, commandId, null);
+            }
         }
 
         protected override VisualStudioHostConfiguration GetVisualStudioHostConfiguration()
